--- a/src/components/menu-bar/menu-bar.dynamic.jsx
+++ b/src/components/menu-bar/menu-bar.dynamic.jsx
@@ -1,0 +1,3 @@
+import {idleRequireWhileLoading} from '../../lib/dynamic-render';
+
+export default idleRequireWhileLoading(() => require('./menu-bar.jsx'));