--- a/src/containers/blocks.dynamic.jsx
+++ b/src/containers/blocks.dynamic.jsx
@@ -1,0 +1,3 @@
+import {idleWhileLoadingAfterFetching} from '../lib/dynamic-render';
+
+export default idleWhileLoadingAfterFetching(() => require('./blocks.jsx'));