--- a/src/containers/sound-tab.dynamic.jsx
+++ b/src/containers/sound-tab.dynamic.jsx
@@ -1,0 +1,5 @@
+import {idleWhileLoading} from '../lib/dynamic-render';
+
+import './sound-tab.jsx';
+
+export default idleWhileLoading(() => require('./sound-tab.jsx'));