--- a/src/containers/target-pane.dynamic.jsx
+++ b/src/containers/target-pane.dynamic.jsx
@@ -1,0 +1,3 @@
+import {idleRequireWhileLoading} from '../lib/dynamic-render';
+
+export default idleRequireWhileLoading(() => require('./target-pane.jsx'));