--- a/src/lib/dynamic-render/index.js
+++ b/src/lib/dynamic-render/index.js
@@ -1,0 +1,32 @@
+export {compose} from 'redux';
+
+export {
+    loadChildren,
+    loadComponent,
+    loadNull
+} from './load-module.jsx';
+
+export {
+    addProps,
+    Null,
+    gate,
+    ifNotReady,
+    ifReady,
+    placeholder
+} from './gate.jsx';
+
+export {
+    schedule
+} from './schedule.jsx';
+
+export {
+    idleWhileLoading,
+    idleRequireWhileLoading,
+    idleWhileLoadingAfterFetching,
+    whileLoading
+} from './definitions.jsx';
+
+export {
+    fetching,
+    loading
+} from './selectors.js';