--- a/src/lib/dynamic-render/schedule.jsx
+++ b/src/lib/dynamic-render/schedule.jsx
@@ -1,4 +1,3 @@
-import PropTypes from 'prop-types';
 import React from 'react';
 
 import {afterReady} from './after-ready.jsx';