--- a/src/lib/dynamic-render/selectors.js
+++ b/src/lib/dynamic-render/selectors.js
@@ -1,0 +1,35 @@
+import {
+    getIsFetchingWithId,
+    getIsLoadingWithId,
+    LoadingState
+} from '../../reducers/project-state';
+
+// Selectors here can provide a descriptive interface for when delay arguments
+// should be which values. If we use only these common functions we can use that
+// as a way to shortcut all of the delays gates. By replacing the functions on
+// Delay with ones that return true.
+
+const loadingState = state => state.scratchGui.projectState.loadingState;
+
+const fetching = state => (
+    loadingState(state) === LoadingState.NOT_LOADED ||
+    getIsFetchingWithId(loadingState(state))
+);
+
+const isLoading = state => getIsLoadingWithId(loadingState(state));
+
+const loadingStateVisible = state => state.scratchGui.modals.loadingProject;
+
+const loading = state => (
+    fetching(state) ||
+    isLoading(state) ||
+    loadingStateVisible(state)
+);
+
+export {
+    fetching,
+    isLoading,
+    loading,
+    loadingState,
+    loadingStateVisible
+};