@mzgoddard 2019-04-08
This PR is architecturally and organizationally ready for review. I'm working on cleaning up remaining lint/test/doc issues.