@mzgoddard 2019-04-10
- Rebase on #4721 
- Fix lint issues