@paulkaplan 2019-04-30
@mzgoddard i'm getting a warning in the console: `Warning: Invalid value for prop `dispatch` on <div> tag. Either remove it from the element, or pass a string or number value to keep it in the DOM.`